+++ Business-boost/src/main/java/com/visumIT/Business/boost/resource/EmployeeResource.java
@@ -3,7 +3,31 @@
  * */
 package com.visumIT.Business.boost.resource;
 
+import java.util.Calendar;
+import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
+
+import org.json.JSONObject;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.visumIT.Business.boost.DTO.EmployeeDTO;
+import com.visumIT.Business.boost.enums.Profile;
 import com.visumIT.Business.boost.exception.ValidationFormat;
 import com.visumIT.Business.boost.functions.ImageValidations;
 import com.visumIT.Business.boost.functions.PartialUpdateValidation;
@@ -13,6 +37,11 @@ import com.visumIT.Business.boost.models.Phone;
 import com.visumIT.Business.boost.repository.CompanyRepository;
 import com.visumIT.Business.boost.repository.EmployeeRepository;
 import com.visumIT.Business.boost.repository.PhoneRepository;
+import com.visumIT.Business.boost.security.UserSS;
+import com.visumIT.Business.boost.services.UserService;
+import com.visumIT.Business.boost.upload.FileUpload;
+import com.visumIT.Business.boost.upload.FileUploadUrl;
+import com.visumIT.Business.boost.upload.FirebaseStorageService;
 import com.visumIT.Business.boost.upload.FileUpload;
 import com.visumIT.Business.boost.upload.FileUploadUrl;
 import com.visumIT.Business.boost.upload.FirebaseStorageService;
@@ -38,6 +67,9 @@ public class EmployeeResource {
 	private EmployeeRepository employeeRepository;
 	
 	@Autowired
+	private BCryptPasswordEncoder bCryptEncoder;
+	
+	@Autowired
 	private FirebaseStorageService firebase;
 
 	@Autowired
@@ -50,9 +82,6 @@ public class EmployeeResource {
 		Company company = new Company();
 		Optional<Company> companyWanted = companyRepository.findById(id);
 		company = company.optionalToCompany(companyWanted);
-		if (!employeeRepository.existsByCompany(company)) {
-			return "404";
-		}
 		if (employeeRepository.existsByEmail(employee.getEmail())) {
 			return "E-mail allready in use";
 
@@ -63,8 +92,14 @@ public class EmployeeResource {
 	}
 
 	// retorna os employees de uma company
+	@PreAuthorize("hasAnyRole('ADMIN')")
 	@GetMapping
 	public ResponseEntity<?> getEmployees(@PathVariable Long id) {
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.ADMIN) || !id.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
+		
 		Optional<Company> companyWanted = companyRepository.findById(id);
 
 		if (companyWanted.isPresent()) {
@@ -78,9 +113,14 @@ public class EmployeeResource {
 	}
 
 	/* Retorna um employee especifico */
+	@PreAuthorize("hasAnyRole('ADMIN')")
 	@GetMapping("{id_employee}")
 	public ResponseEntity<?> getEmployee(@PathVariable(name = "id") Long id_company,
 			@PathVariable(name = "id_employee") Long id_employee) {
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.ADMIN) || !id_company.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 		Optional<Employee> employeeWanted = employeeRepository.findById(id_employee);
 		Optional<Company> companyWanted = companyRepository.findById(id_company);
 
@@ -94,20 +134,29 @@ public class EmployeeResource {
 	}
 
 	// cadastro de um novo employee
+	@PreAuthorize("hasAnyRole('ADMIN')")
 	@PostMapping
 	public ResponseEntity<?> saveEmployee(@Valid @RequestBody Employee employee,
 			@PathVariable(name = "id") Long id_company, BindingResult bindingResult) {
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.ADMIN) || !id_company.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
+		
 		String message = validEmployee(id_company, employee);
+
 		if (!message.equals("ok")) {
 			if (message.equals("404")) {
 				return ResponseEntity.badRequest().build();
 			}
-			return ResponseEntity.badRequest().body(new JSONObject().put("message", message.toString()));
+			return ResponseEntity.badRequest().body(new JSONObject().put("message", message).toString());
 		}
+
 		if (bindingResult.hasErrors()) {
 			return ResponseEntity.badRequest().body(ValidationFormat.formatarErros(bindingResult));
 
 		} else {
+			employee.setPassword(bCryptEncoder.encode(employee.getPassword()));
 			employeeRepository.save(employee);
 			Optional<Company> companyOptional = companyRepository.findById(id_company);
 
@@ -118,17 +167,25 @@ public class EmployeeResource {
 				tel.setEmployee(employee);
 				phoneRepository.save(tel);
 			}
+			System.out.println("----------------------");
+			System.out.println(employee.getProfiles());
 			employeeRepository.save(employee);
-			return ResponseEntity.status(HttpStatus.CREATED).body(employee);
+			EmployeeDTO dto = new EmployeeDTO();
+			return ResponseEntity.status(HttpStatus.CREATED).body(dto.toEmployeeDTO(employee));
 
 		}
 	}
 
 	// update parcial
+	@PreAuthorize("hasAnyRole('ADMIN')")
 	@PatchMapping("/{id_employee}")
 	public ResponseEntity<?> partialUpdate(@Valid @RequestBody Employee bodyEmployee,
 			@PathVariable(name = "id_employee") Long id_employee, @PathVariable(name = "id") Long id_company) 
 					throws IllegalAccessException {
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.ADMIN) || !id_company.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 		
 		Optional<Company> companyWanted = companyRepository.findById(id_company);
 		Company company = new Company();
@@ -148,10 +205,16 @@ public class EmployeeResource {
 	}
 
 	/* update completo */
+	@PreAuthorize("hasAnyRole('ADMIN')")
 	@PutMapping("/{id_employee}")
 	public ResponseEntity<?> fullEmployeeUpdate(@Valid @RequestBody Employee employee,
 			@PathVariable(name = "id") Long id_company, @PathVariable(name = "id_employee") Long id_employee,
 			BindingResult bindingResult) {
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.ADMIN) || !id_company.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
+		
 
 		Optional<Company> companyWanted = companyRepository.findById(id_company);
 		Optional<Employee> employeeWanted = employeeRepository.findById(id_employee);
@@ -180,8 +243,14 @@ public class EmployeeResource {
 	}
 
 	/* deletar employee */
+	@PreAuthorize("hasAnyRole('ADMIN')")
 	@DeleteMapping("/{id_employee}")
-	public ResponseEntity<?> deletarEmployee(@PathVariable Long id_employee) {
+	public ResponseEntity<?> deletarEmployee(@PathVariable Long id_employee, @PathVariable(name="id") Long id) {
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.ADMIN) || !id.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
+		
 		if (employeeRepository.existsById(id_employee)) {
 			employeeRepository.deleteById(id_employee);
 			return ResponseEntity.noContent().build();
@@ -191,9 +260,16 @@ public class EmployeeResource {
 	}
 
 	/*upload de foto*/
+	@PreAuthorize("hasAnyRole('EMPLOYEE')")
 	@PatchMapping("/{id_employee}/photos")
 	public ResponseEntity<?>updatePhotograph (@RequestBody FileUpload file,
 			@PathVariable(name = "id_employee") Long id_employee, @PathVariable(name = "id") Long id_company) {
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.EMPLOYEE) || !id_employee.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
+		
+		
 		ImageValidations imageValidations = new ImageValidations();
 		Company company = new Company();
 		company = company.optionalToCompany(companyRepository.findById(id_company));
@@ -223,8 +299,13 @@ public class EmployeeResource {
 	}
 	
 	/*deletar foto*/
+	@PreAuthorize("hasAnyRole('EMPLOYEE')")
 	@DeleteMapping("{id_employee}/photos")
 	public ResponseEntity<?> deleteLogo(@PathVariable(name = "id_employee") Long id_employee, @PathVariable(name = "id") Long id_company){
+		UserSS user = UserService.authenticated();
+		if (user == null || !user.hasRole(Profile.EMPLOYEE) || !id_employee.equals(user.getId())) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+		}
 		Company company = new Company();
 		company = company.optionalToCompany(companyRepository.findById(id_company));
 		if (employeeRepository.existsByCompany(company)) {
